CC HW3/backend/requirements.txt
Flask
gunicorn
google-cloud-datastore
google-cloud-storage
flask-cors
